Added support for genes as well as filter for genomeBrowser
